@@ -26,24 +26,30 @@
         }
     ],
     "require": {
-        "php": "^7.3 || ^8.0",
+        "php": "^7.4 || ~8.0",
         "laminas/laminas-eventmanager": "^3.3",
         "laminas/laminas-router": "^3.4",
         "laminas/laminas-servicemanager": "^3.6",
         "laminas/laminas-stdlib": "^3.3",
-        "slm/queue": "3.0-beta.3",
+        "slm/queue": "^3.0",
         "doctrine/dbal": "^2.12",
         "doctrine/annotations": "^1.8",
         "laminas/laminas-config": "^3.4",
         "laminas/laminas-cli": "^1.1.1"
+        
     },
     "require-dev": {
+        "ext-sqlite3": "*",
+        "doctrine/doctrine-orm-module": "^4.0",
         "laminas/laminas-modulemanager": "^2.10",
         "laminas/laminas-view": "^2.12",
         "laminas/laminas-log": "^2.13",
         "laminas/laminas-i18n": "^2.11",
         "laminas/laminas-serializer": "^2.10",
-        "phpunit/phpunit": "^9"
+        "laminas/laminas-mvc": "^3.3",
+        "doctrine/orm": "^2.10",
+        "phpunit/phpunit": "^9",
+        "squizlabs/php_codesniffer": "3.*"
     },
     "suggest": {
         "doctrine/doctrine-orm-module": "If you use Doctrine in combination with Laminas",
@@ -70,5 +76,14 @@
         "psr-4": {
             "SlmQueueDoctrineTest\\": "tests/src"
         }
+    },
+    "scripts": {
+        "cs-check": "phpcs",
+        "cs-fix": "phpcbf",
+        "test": [
+            "phpunit",
+            "@composer test --working-dir=tests/laminas-runner",
+            "@composer test --working-dir=tests/mezzio-runner"
+        ]
     }
 }
