@@ -24,9 +24,14 @@ class DoctrineQueueFactory implements FactoryInterface
         $queueOptions  = new DoctrineOptions($options);
 
         /** @var $connection \Doctrine\DBAL\Connection */
-        $connection       = $container->get($queueOptions->getConnection());
-        $jobPluginManager = $container->get(JobPluginManager::class);
+        $connection          = $container->get($queueOptions->getConnection());
+        $jobPluginManager    = $container->get(JobPluginManager::class);
 
-        return new DoctrineQueue($connection, $queueOptions, $requestedName, $jobPluginManager);
+        return new DoctrineQueue(
+            $connection,
+            $queueOptions,
+            $requestedName,
+            $jobPluginManager
+        );
     }
 }
