@@ -1,6 +1,6 @@
 <?php
 
-namespace TestModule;
+namespace App;
 
 use SlmQueue\Job\AbstractJob;
 
@@ -8,6 +8,6 @@ class TestJob extends AbstractJob
 {
     public function execute()
     {
-        file_put_contents('temp/succesfull', 'YES');
+        file_put_contents('temp/succesful', 'YES');
     }
 }
